+++ packages/kit/src/views/Onboardingv2/pages/ConnectYourDevice.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
-import { EDeviceType, HardwareErrorCode } from '@onekeyfe/hd-shared';
+import { EDeviceType } from '@onekeyfe/hd-shared';
 import { useIsFocused } from '@react-navigation/core';
-import { get, isString } from 'lodash';
+import { isString } from 'lodash';
 import natsort from 'natsort';
 import { useIntl } from 'react-intl';
 import { Linking, StyleSheet } from 'react-native';
@@ -31,7 +31,6 @@ import {
   usePopoverContext,
 } from '@onekeyhq/components';
 import { usePromptWebDeviceAccess } from '@onekeyhq/kit/src/hooks/usePromptWebDeviceAccess';
-import type { IDBCreateHwWalletParamsBase } from '@onekeyhq/kit-bg/src/dbs/local/types';
 import { useSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { HARDWARE_BRIDGE_DOWNLOAD_URL } from '@onekeyhq/shared/src/config/appConfig';
 import {
@@ -45,21 +44,12 @@ import {
   NeedBluetoothPermissions,
   NeedBluetoothTurnedOn,
   NeedOneKeyBridge,
-  OneKeyHardwareError,
-  OneKeyLocalError,
 } from '@onekeyhq/shared/src/errors';
 import { convertDeviceError } from '@onekeyhq/shared/src/errors/utils/deviceErrorUtils';
-import errorToastUtils from '@onekeyhq/shared/src/errors/utils/errorToastUtils';
-import {
-  EAppEventBusNames,
-  appEventBus,
-} from '@onekeyhq/shared/src/eventBus/appEventBus';
 import bleManagerInstance from '@onekeyhq/shared/src/hardware/bleManager';
 import { checkBLEPermissions } from '@onekeyhq/shared/src/hardware/blePermissions';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
-import { defaultLogger } from '@onekeyhq/shared/src/logger/logger';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import { EOnboardingPages } from '@onekeyhq/shared/src/routes/onboarding';
 import type { IOnboardingParamListV2 } from '@onekeyhq/shared/src/routes/onboardingv2';
 import { EOnboardingPagesV2 } from '@onekeyhq/shared/src/routes/onboardingv2';
 import {
@@ -73,11 +63,7 @@ import {
   EHardwareTransportType,
 } from '@onekeyhq/shared/types';
 import { EConnectDeviceChannel } from '@onekeyhq/shared/types/connectDevice';
-import { EOneKeyDeviceMode } from '@onekeyhq/shared/types/device';
-import type {
-  IConnectYourDeviceItem,
-  IOneKeyDeviceFeatures,
-} from '@onekeyhq/shared/types/device';
+import type { IConnectYourDeviceItem } from '@onekeyhq/shared/types/device';
 
 import backgroundApiProxy from '../../../background/instance/backgroundApiProxy';
 import { AccountSelectorProviderMirror } from '../../../components/AccountSelector/AccountSelectorProvider';
@@ -90,20 +76,13 @@ import { HyperlinkText } from '../../../components/HyperlinkText';
 import { ListItem } from '../../../components/ListItem';
 import { WalletAvatar } from '../../../components/WalletAvatar';
 import useAppNavigation from '../../../hooks/useAppNavigation';
-import { useHelpLink } from '../../../hooks/useHelpLink';
 import { useThemeVariant } from '../../../hooks/useThemeVariant';
-import { useUserWalletProfile } from '../../../hooks/useUserWalletProfile';
-import { useAccountSelectorActions } from '../../../states/jotai/contexts/accountSelector';
-import { useFirmwareUpdateActions } from '../../FirmwareUpdate/hooks/useFirmwareUpdateActions';
-import { useFirmwareVerifyDialog } from '../../Onboarding/pages/ConnectHardwareWallet/FirmwareVerifyDialog';
-import { useSelectAddWalletTypeDialog } from '../../Onboarding/pages/ConnectHardwareWallet/SelectAddWalletTypeDialog';
 import { OnboardingLayout } from '../components/OnboardingLayout';
+import { useDeviceConnect } from '../hooks/useDeviceConnect';
 import {
   getDeviceLabel,
   getForceTransportType,
-  getHardwareCommunicationTypeString,
   sortDevicesData,
-  trackHardwareWalletConnection,
 } from '../utils';
 
 import type { IDeviceType, SearchDevice } from '@onekeyfe/hd-core';
@@ -132,22 +111,17 @@ function BridgeNotInstalledDialogContent(_props: { error: NeedOneKeyBridge }) {
 }
 
 interface IDeviceConnectionProps {
+  tabValue: EConnectDeviceChannel;
   deviceTypeItems: EDeviceType[];
   onDeviceConnect: (device: SearchDevice) => Promise<void>;
-  onSelectAddWalletType: (params: {
-    device: SearchDevice;
-    isFirmwareVerified: boolean;
-  }) => Promise<void>;
 }
 
 // Common device list and connection logic
 function useDeviceConnection({
   tabValue,
-  onDeviceConnect,
-  onSelectAddWalletType,
 }: {
   tabValue: EConnectDeviceChannel;
-} & IDeviceConnectionProps) {
+}) {
   const intl = useIntl();
   const [connectStatus, setConnectStatus] = useState(EConnectionStatus.init);
   const [searchedDevices, setSearchedDevices] = useState<SearchDevice[]>([]);
@@ -354,20 +328,8 @@ function useDeviceConnection({
         title: item.name,
         src: HwWalletAvatarImages[getDeviceAvatarImage(item.deviceType)],
         device: item,
-        onFirmwareVerified: async () => {
-          // Ensure device scanning is completely stopped before connecting
-          await ensureStopScan();
-          await onDeviceConnect(item);
-        },
-        onCreateWallet: async () => {
-          await ensureStopScan();
-          await onSelectAddWalletType({
-            device: item,
-            isFirmwareVerified: true,
-          });
-        },
       })),
-    [searchedDevices, onSelectAddWalletType, ensureStopScan, onDeviceConnect],
+    [searchedDevices],
   );
 
   return {
@@ -723,11 +685,8 @@ function DeviceVideo({
 function USBOrBLEConnectionIndicator({
   tabValue,
   onDeviceConnect,
-  onSelectAddWalletType,
   deviceTypeItems,
-}: {
-  tabValue: EConnectDeviceChannel;
-} & IDeviceConnectionProps) {
+}: IDeviceConnectionProps) {
   const themeVariant = useThemeVariant();
   const intl = useIntl();
   const navigation = useAppNavigation();
@@ -737,9 +696,6 @@ function USBOrBLEConnectionIndicator({
   // Use the shared device connection logic
   const deviceConnection = useDeviceConnection({
     tabValue,
-    deviceTypeItems,
-    onDeviceConnect,
-    onSelectAddWalletType,
   });
 
   const {
@@ -935,6 +891,7 @@ function USBOrBLEConnectionIndicator({
                     onPress={() => {
                       navigation.push(EOnboardingPagesV2.CheckAndUpdate, {
                         deviceData: data,
+                        tabValue,
                       });
                     }}
                     userSelect="none"
@@ -958,7 +915,7 @@ function USBOrBLEConnectionIndicator({
 function BluetoothConnectionIndicator({
   deviceTypeItems,
   onDeviceConnect,
-  onSelectAddWalletType,
+  tabValue,
 }: IDeviceConnectionProps) {
   const intl = useIntl();
   const isFocused = useIsFocused();
@@ -1079,10 +1036,7 @@ function BluetoothConnectionIndicator({
 
   // Use shared device connection logic for Bluetooth
   const deviceConnection = useDeviceConnection({
-    tabValue: EConnectDeviceChannel.bluetooth,
-    deviceTypeItems,
-    onDeviceConnect: handleBluetoothDeviceConnect,
-    onSelectAddWalletType,
+    tabValue,
   });
 
   const { devicesData, scanDevice, stopScan } = deviceConnection;
@@ -1220,7 +1174,10 @@ function BluetoothConnectionIndicator({
                     key={device.device?.connectId}
                     drillIn
                     onPress={() => {
-                      navigation.push(EOnboardingPagesV2.CheckAndUpdate);
+                      navigation.push(EOnboardingPagesV2.CheckAndUpdate, {
+                        deviceData: device,
+                        tabValue,
+                      });
                     }}
                     userSelect="none"
                   >
@@ -1312,547 +1269,13 @@ function ConnectYourDevicePage({
   }, [intl]);
   const [tabValue, setTabValue] = useState(tabOptions[0]?.value);
 
-  const { gtMd } = useMedia();
-  const navigation = useAppNavigation();
-  const actions = useAccountSelectorActions();
-  const { showFirmwareVerifyDialog } = useFirmwareVerifyDialog();
-  const { showSelectAddWalletTypeDialog } = useSelectAddWalletTypeDialog();
-  const fwUpdateActions = useFirmwareUpdateActions();
-  const { isSoftwareWalletOnlyUser } = useUserWalletProfile();
-  const [{ hardwareTransportType }] = useSettingsPersistAtom();
-  const [isCheckingDeviceLoading, setIsChecking] = useState(false);
+  const { onDeviceConnect } = useDeviceConnect();
 
-  const handleRestoreWalletPress = useCallback(
-    ({ deviceType }: { deviceType: IDeviceType }) => {
-      navigation.push(EOnboardingPages.ActivateDevice, {
-        tutorialType: 'restore',
-        deviceType,
-      });
-    },
-    [navigation],
-  );
-
-  const handleSetupNewWalletPress = useCallback(
-    ({ deviceType }: { deviceType: IDeviceType }) => {
-      navigation.push(EOnboardingPages.ActivateDevice, {
-        tutorialType: 'create',
-        deviceType,
-      });
-    },
-    [navigation],
-  );
-
-  const connectDevice = useCallback(async (device: SearchDevice) => {
-    try {
-      return await backgroundApiProxy.serviceHardware.connect({
-        device,
-      });
-    } catch (error: any) {
-      if (error instanceof OneKeyHardwareError) {
-        const { code, message } = error;
-        if (
-          code === HardwareErrorCode.CallMethodNeedUpgradeFirmware ||
-          code === HardwareErrorCode.BlePermissionError ||
-          code === HardwareErrorCode.BleLocationError
-        ) {
-          return;
-        }
-        Toast.error({
-          title: message || 'DeviceConnectError',
-        });
-      } else {
-        console.error('connectDevice error:', get(error, 'message', ''));
-      }
-    }
-  }, []);
-
-  const requestsUrl = useHelpLink({ path: 'requests/new' });
-
-  const handleNotActivatedDevicePress = useCallback(
-    ({ deviceType }: { deviceType: IDeviceType }) => {
-      const dialog = Dialog.show({
-        icon: 'WalletCryptoOutline',
-        title: intl.formatMessage({
-          id: ETranslations.onboarding_activate_device,
-        }),
-        description: intl.formatMessage({
-          id: ETranslations.onboarding_activate_device_help_text,
-        }),
-        dismissOnOverlayPress: false,
-        renderContent: (
-          <Stack>
-            <ListItem
-              alignItems="flex-start"
-              icon="PlusCircleOutline"
-              title={intl.formatMessage({
-                id: ETranslations.onboarding_activate_device_by_set_up_new_wallet,
-              })}
-              subtitle={intl.formatMessage({
-                id: ETranslations.onboarding_activate_device_by_set_up_new_wallet_help_text,
-              })}
-              drillIn
-              onPress={async () => {
-                await dialog.close();
-                handleSetupNewWalletPress({ deviceType });
-              }}
-              borderWidth={StyleSheet.hairlineWidth}
-              borderColor="$borderSubdued"
-              m="$0"
-              py="$2.5"
-              bg="$bgSubdued"
-            />
-            <ListItem
-              alignItems="flex-start"
-              icon="ArrowBottomCircleOutline"
-              title={intl.formatMessage({
-                id: ETranslations.onboarding_activate_device_by_restore,
-              })}
-              subtitle={intl.formatMessage({
-                id: ETranslations.onboarding_activate_device_by_restore_help_text,
-              })}
-              drillIn
-              onPress={async () => {
-                await dialog.close();
-                const packageAlertDialog = Dialog.show({
-                  tone: 'warning',
-                  icon: 'PackageDeliveryOutline',
-                  title: intl.formatMessage({
-                    id: ETranslations.onboarding_activate_device_by_restore_warning,
-                  }),
-                  dismissOnOverlayPress: false,
-                  description: intl.formatMessage({
-                    id: ETranslations.onboarding_activate_device_by_restore_warning_help_text,
-                  }),
-                  showFooter: false,
-                  renderContent: (
-                    <XStack gap="$2.5">
-                      <Button
-                        flex={1}
-                        size="large"
-                        $gtMd={{ size: 'medium' } as any}
-                        onPress={() => Linking.openURL(requestsUrl)}
-                      >
-                        {intl.formatMessage({
-                          id: ETranslations.global_contact_us,
-                        })}
-                      </Button>
-                      <Button
-                        flex={1}
-                        variant="primary"
-                        size="large"
-                        $gtMd={{ size: 'medium' } as any}
-                        onPress={async () => {
-                          await packageAlertDialog.close();
-                          handleRestoreWalletPress({ deviceType });
-                        }}
-                      >
-                        {intl.formatMessage({
-                          id: ETranslations.global_continue,
-                        })}
-                      </Button>
-                    </XStack>
-                  ),
-                });
-              }}
-              borderWidth={StyleSheet.hairlineWidth}
-              borderColor="$borderSubdued"
-              m="$0"
-              mt="$2.5"
-              py="$2.5"
-              bg="$bgSubdued"
-            />
-          </Stack>
-        ),
-        showFooter: false,
-      });
-    },
-    [handleRestoreWalletPress, handleSetupNewWalletPress, intl, requestsUrl],
-  );
-
-  const extractDeviceState = useCallback(
-    (features: IOneKeyDeviceFeatures) => ({
-      unlockedAttachPin: features.unlocked_attach_pin,
-      unlocked: features.unlocked,
-      passphraseEnabled: Boolean(features.passphrase_protection),
-      deviceId: features.device_id,
-    }),
-    [],
-  );
-
-  const closeDialogAndReturn = useCallback(
-    async (device: SearchDevice, options: { skipDelayClose?: boolean }) => {
-      setIsChecking(false);
-      void backgroundApiProxy.serviceHardwareUI.closeHardwareUiStateDialog({
-        connectId: device.connectId ?? '',
-        hardClose: true,
-        skipDelayClose: options.skipDelayClose,
-      });
-    },
-    [],
-  );
-
-  type IWalletCreationStrategy = {
-    createHiddenWalletOnly: boolean;
-    createStandardWalletOnly: boolean;
-  };
-
-  const determineWalletCreationStrategy = useCallback(
-    async (
-      deviceState: ReturnType<typeof extractDeviceState>,
-      device: SearchDevice,
-    ): Promise<IWalletCreationStrategy | null> => {
-      if (!deviceState.unlocked) {
-        return {
-          createHiddenWalletOnly: false,
-          createStandardWalletOnly: true,
-        };
-      }
-
-      if (deviceState.unlockedAttachPin) {
-        return {
-          createHiddenWalletOnly: deviceState.passphraseEnabled,
-          createStandardWalletOnly: !deviceState.passphraseEnabled,
-        };
-      }
-
-      const existsStandardWallet =
-        await backgroundApiProxy.serviceAccount.existsHwStandardWallet({
-          connectId: device.connectId ?? '',
-          deviceId: deviceState.deviceId ?? '',
-        });
-
-      if (existsStandardWallet) {
-        return {
-          createHiddenWalletOnly: deviceState.passphraseEnabled,
-          createStandardWalletOnly: !deviceState.passphraseEnabled,
-        };
-      }
-
-      if (!deviceState.passphraseEnabled) {
-        return {
-          createHiddenWalletOnly: false,
-          createStandardWalletOnly: true,
-        };
-      }
-
-      const walletType = await showSelectAddWalletTypeDialog();
-      if (walletType === 'Standard') {
-        return {
-          createHiddenWalletOnly: false,
-          createStandardWalletOnly: true,
-        };
-      }
-      if (walletType === 'Hidden') {
-        return {
-          createHiddenWalletOnly: true,
-          createStandardWalletOnly: false,
-        };
-      }
-
-      return null;
-    },
-    [showSelectAddWalletTypeDialog],
-  );
-
-  const createHwWallet = useCallback(
-    async (
-      device: SearchDevice,
-      strategy: IWalletCreationStrategy,
-      features: IOneKeyDeviceFeatures,
-      isFirmwareVerified?: boolean,
-      deviceState?: ReturnType<typeof extractDeviceState>,
-    ) => {
-      try {
-        navigation.push(EOnboardingPages.FinalizeWalletSetup);
-
-        const params: IDBCreateHwWalletParamsBase = {
-          device,
-          hideCheckingDeviceLoading: true,
-          features,
-          isFirmwareVerified,
-          defaultIsTemp: true,
-          isAttachPinMode: deviceState?.unlockedAttachPin,
-        };
-        if (strategy.createStandardWalletOnly) {
-          await actions.current.createHWWalletWithoutHidden(params);
-        } else {
-          await actions.current.createHWWalletWithHidden(params);
-        }
-
-        await trackHardwareWalletConnection({
-          status: 'success',
-          deviceType: device.deviceType,
-          features,
-          hardwareTransportType,
-          isSoftwareWalletOnlyUser,
-        });
-
-        await actions.current.updateHwWalletsDeprecatedStatus({
-          connectId: device.connectId ?? '',
-          deviceId: features.device_id || device.deviceId || '',
-        });
-      } catch (error) {
-        errorToastUtils.toastIfError(error);
-        navigation.pop();
-        await trackHardwareWalletConnection({
-          status: 'failure',
-          deviceType: device.deviceType,
-          features,
-          hardwareTransportType,
-          isSoftwareWalletOnlyUser,
-        });
-        throw error;
-      } finally {
-        await closeDialogAndReturn(device, { skipDelayClose: false });
-      }
-    },
-    [
-      actions,
-      closeDialogAndReturn,
-      hardwareTransportType,
-      isSoftwareWalletOnlyUser,
-      navigation,
-    ],
-  );
-
-  const selectAddWalletType = useCallback(
-    async ({
-      device,
-      isFirmwareVerified,
-    }: {
-      device: SearchDevice;
-      isFirmwareVerified?: boolean;
-    }) => {
-      setIsChecking(true);
-
-      void backgroundApiProxy.serviceHardwareUI.showDeviceProcessLoadingDialog({
-        connectId: device.connectId ?? '',
-      });
-
-      let features: IOneKeyDeviceFeatures | undefined;
-
-      try {
-        features =
-          await backgroundApiProxy.serviceHardware.getFeaturesWithUnlock({
-            connectId: device.connectId ?? '',
-          });
-      } catch (error) {
-        await closeDialogAndReturn(device, { skipDelayClose: true });
-        return;
-      }
-
-      const deviceState = extractDeviceState(features);
-      const strategy = await determineWalletCreationStrategy(
-        deviceState,
-        device,
-      );
-
-      console.log('Current hardware wallet State', deviceState, strategy);
-      if (!strategy) {
-        await closeDialogAndReturn(device, { skipDelayClose: true });
-        return;
-      }
-
-      await createHwWallet(
-        device,
-        strategy,
-        features,
-        isFirmwareVerified,
-        deviceState,
-      );
-    },
-    [
-      extractDeviceState,
-      determineWalletCreationStrategy,
-      createHwWallet,
-      closeDialogAndReturn,
-    ],
-  );
-
-  // Shared device connection handler
   const handleDeviceConnect = useCallback(
     async (device: SearchDevice) => {
-      // Ensure all scanning and polling activities are stopped before connecting
-      console.log('handleDeviceConnect: Starting device connection process');
-
-      defaultLogger.account.wallet.addWalletStarted({
-        addMethod: 'ConnectHWWallet',
-        details: {
-          hardwareWalletType: 'Standard',
-          communication: getHardwareCommunicationTypeString(
-            hardwareTransportType,
-          ),
-        },
-        isSoftwareWalletOnlyUser,
-      });
-
-      if (device.deviceType === 'unknown') {
-        Toast.error({
-          title: intl.formatMessage({
-            id: ETranslations.hardware_connect_unknown_device_error,
-          }),
-        });
-        return;
-      }
-
-      try {
-        void backgroundApiProxy.serviceHardwareUI.showCheckingDeviceDialog({
-          connectId: device.connectId ?? '',
-        });
-
-        const handleBootloaderMode = (existsFirmware: boolean) => {
-          fwUpdateActions.showBootloaderMode({
-            connectId: device.connectId ?? undefined,
-            existsFirmware,
-          });
-          console.log('Device is in bootloader mode', device);
-          throw new OneKeyLocalError('Device is in bootloader mode');
-        };
-
-        if (
-          await deviceUtils.isBootloaderModeFromSearchDevice({
-            device: device as any,
-          })
-        ) {
-          const existsFirmware =
-            await deviceUtils.existsFirmwareFromSearchDevice({
-              device: device as any,
-            });
-          handleBootloaderMode(existsFirmware);
-          return;
-        }
-
-        // Set global transport type based on selected channel before connecting
-        let forceTransportType: EHardwareTransportType | undefined;
-        if (tabValue === EConnectDeviceChannel.bluetooth) {
-          forceTransportType = EHardwareTransportType.DesktopWebBle;
-        } else {
-          forceTransportType = await getForceTransportType(tabValue);
-        }
-        if (forceTransportType) {
-          await backgroundApiProxy.serviceHardware.setForceTransportType({
-            forceTransportType,
-          });
-        }
-
-        const features = await connectDevice(device);
-
-        if (!features) {
-          await trackHardwareWalletConnection({
-            status: 'failure',
-            isSoftwareWalletOnlyUser,
-            deviceType: device.deviceType,
-            features,
-            hardwareTransportType: forceTransportType || hardwareTransportType,
-          });
-          throw new OneKeyHardwareError(
-            'connect device failed, no features returned',
-          );
-        }
-
-        if (await deviceUtils.isBootloaderModeByFeatures({ features })) {
-          const existsFirmware = await deviceUtils.existsFirmwareByFeatures({
-            features,
-          });
-          handleBootloaderMode(existsFirmware);
-          return;
-        }
-
-        let deviceType = await deviceUtils.getDeviceTypeFromFeatures({
-          features,
-        });
-        if (deviceType === 'unknown') {
-          deviceType = device.deviceType || deviceType;
-        }
-
-        const deviceMode = await deviceUtils.getDeviceModeFromFeatures({
-          features,
-        });
-
-        if (deviceMode === EOneKeyDeviceMode.backupMode) {
-          await trackHardwareWalletConnection({
-            status: 'failure',
-            deviceType,
-            isSoftwareWalletOnlyUser,
-            features,
-            hardwareTransportType: forceTransportType || hardwareTransportType,
-          });
-          Toast.error({
-            title: 'Device is in backup mode',
-          });
-          return;
-        }
-
-        const shouldAuthenticateFirmware =
-          await backgroundApiProxy.serviceHardware.shouldAuthenticateFirmware({
-            device: {
-              ...device,
-              deviceId: device.deviceId || features.device_id,
-            },
-          });
-
-        if (shouldAuthenticateFirmware) {
-          void backgroundApiProxy.serviceHardwareUI.closeHardwareUiStateDialog({
-            connectId: device.connectId ?? '',
-            hardClose: false,
-            skipDelayClose: true,
-            deviceResetToHome: false,
-          });
-          await showFirmwareVerifyDialog({
-            device,
-            features,
-            onContinue: async ({ checked }: { checked: boolean }) => {
-              setIsChecking(false);
-              if (deviceMode === EOneKeyDeviceMode.notInitialized) {
-                handleNotActivatedDevicePress({ deviceType });
-              }
-            },
-            onClose: () => {
-              setIsChecking(false);
-            },
-          });
-          return;
-        }
-        void backgroundApiProxy.serviceHardwareUI.closeHardwareUiStateDialog({
-          connectId: device.connectId ?? '',
-          hardClose: false,
-          skipDelayClose: true,
-          deviceResetToHome: false,
-        });
-
-        if (deviceMode === EOneKeyDeviceMode.notInitialized) {
-          handleNotActivatedDevicePress({ deviceType });
-        }
-
-        appEventBus.emit(EAppEventBusNames.EmitFirmwareVerifyResult, {
-          verified: true,
-          device,
-          payload: {
-            deviceType: device.deviceType,
-            data: '',
-            cert: '',
-            signature: '',
-          },
-          result: undefined,
-        });
-      } catch (error) {
-        // Clear force transport type on device connection error
-        void backgroundApiProxy.serviceHardware.clearForceTransportType();
-        void backgroundApiProxy.serviceHardwareUI.cleanHardwareUiState();
-        console.error('handleDeviceConnect error:', error);
-        throw error;
-      }
+      await onDeviceConnect(device, tabValue);
     },
-    [
-      tabValue,
-      hardwareTransportType,
-      isSoftwareWalletOnlyUser,
-      intl,
-      fwUpdateActions,
-      connectDevice,
-      showFirmwareVerifyDialog,
-      handleNotActivatedDevicePress,
-    ],
+    [onDeviceConnect, tabValue],
   );
 
   return (
@@ -1887,14 +1310,13 @@ function ConnectYourDevicePage({
                 tabValue={tabValue}
                 deviceTypeItems={deviceTypeItems}
                 onDeviceConnect={handleDeviceConnect}
-                onSelectAddWalletType={selectAddWalletType}
               />
             ) : null}
             {tabValue === EConnectDeviceChannel.bluetooth ? (
               <BluetoothConnectionIndicator
+                tabValue={tabValue}
                 deviceTypeItems={deviceTypeItems}
                 onDeviceConnect={handleDeviceConnect}
-                onSelectAddWalletType={selectAddWalletType}
               />
             ) : null}
           </OnboardingLayout.ConstrainedContent>
